Title: 최솟값 3개, Time: 185ms, Memory: 1MB, Status: Passed - CodeTree

+++ 240409/최솟값 3개/top-3-smallest-number.cpp
@@ -19,7 +19,7 @@ int main() {
             cout << -1 << endl;
         
         else{
-            int output = 1;
+            long long output = 1;
             for(int i=0; i<3; i++){
                 tmp = -pq.top();
                 output *= tmp;
